requirements.txt
great_expectations==1.3.7
pytest==8.3.4
faker==19.3.0
delta-spark==2.3.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
mypy==0.991
flake8==7.1.2
black==25.1.0
isort==6.0.0
pyspark==3.3.2